+++ NEWS.md
@@ -1,11 +1,17 @@
-# JointFPM 1.0.1
+# JointFPM 1.1.0
 
-# Version 1.0.1
+- Added standardization for estimating marginal estimates of the mean number of events and differences thereof
+- Added checks of user inputs to `JointFPM()`
+- Added checks of user inputs to `predict.JointFPM()`
+- Improved error messages
+- Improved testing coverage
+
+# JointFPM 1.0.1
 
 - Fixed bug when calculating difference between two mean number of events functions
 - Added tests for predict function
 
-# Version 1.0.0
+# JointFPM 1.0.0
 
 - First stable release